ref: Rename module 'bucketing' to 'aggregation'
